requirements.txt
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
nexmo==2.4.0
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.11
twilio==6.33.1
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
